+++ micro/src/org/cougaar/microedition/tini/TiniRotationAndLimitSensor.java
@@ -164,8 +164,8 @@ public class TiniRotationAndLimitSensor extends RotationAndLimitSensorResource {
     }
   }
 
-  public double getValue() {
-    return 0.0;
+  public long getValue() {
+    return 0;
   }
 
   public void setUnits(String newUnits) {
